backtest_ui/requirements.txt
alembic==1.8.1
fastapi==0.85.1
fastapi-sqlalchemy==0.2.1
psycopg2==2.9.4
pydantic==1.10.2 
python-dotenv==0.21.0
uvicorn==0.19.0